Refactor to tokens instead of ast
